projects/ToDoList-FastAPI-ElectronVue/pyproject.toml
[tool.poetry]
name = "todolist-fastapi-electronvue"
version = "0.1.0"
description = ""
authors = ["Kazaz-Or <or.kazaz@dazn.com>"]
readme = "README.md"
packages = [{include = "todolist_fastapi_electronvue"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.95.1"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
